+++ library/Imbo/Model/Group.php
@@ -13,7 +13,63 @@ namespace Imbo\Model;
 /**
  * Group model
  *
- * @author Espen Hovlandsdal <espen@hovlandsdal.com>
+ * @author Christer Edvartsen <cogo@starzinger.net>
  * @package Models
  */
-class Group extends ArrayModel {}
+class Group implements ModelInterface {
+    /**
+     * Name of the group
+     *
+     * @var string
+     */
+    private $name;
+
+    /**
+     * Resources
+     *
+     * @var string[]
+     */
+    private $resources = [];
+
+    /**
+     * Set the group name
+     *
+     * @param string $name The name of the group
+     * @return self
+     */
+    public function setName($name) {
+        $this->name = $name;
+
+        return $this;
+    }
+
+    /**
+     * Get the group name
+     *
+     * @return string
+     */
+    public function getName() {
+        return $this->name;
+    }
+
+    /**
+     * Set the group resources
+     *
+     * @param string[] $resources
+     * @return self
+     */
+    public function setResources(array $resources = []) {
+        $this->resources = $resources;
+
+        return $this;
+    }
+
+    /**
+     * Get the group resources
+     *
+     * @return string[]
+     */
+    public function getResources() {
+        return $this->resources;
+    }
+}