+++ tests/phpunit/ImboIntegrationTest/EventListener/ImageVariations/Database/MongoTest.php
@@ -1,0 +1,59 @@
+<?php
+/**
+ * This file is part of the Imbo package
+ *
+ * (c) Christer Edvartsen <cogo@starzinger.net>
+ *
+ * For the full copyright and license information, please view the LICENSE file that was
+ * distributed with this source code.
+ */
+
+namespace ImboIntegrationTest\EventListener\ImageVariations\Database;
+
+use Imbo\EventListener\ImageVariations\Database\Mongo,
+    MongoDB\Client as MongoClient;
+
+/**
+ * @covers Imbo\EventListener\ImageVariations\Database\Mongo
+ * @group integration
+ * @group database
+ * @group mongo
+ */
+class MongoTest extends DatabaseTests {
+    private $databaseName = 'imboIntegrationTestDatabase';
+
+    /**
+     * @see ImboIntegrationTest\EventListener\ImageVariations\Database\DatabaseTests::getAdapter()
+     */
+    protected function getAdapter() {
+        return new Mongo([
+            'databaseName' => $this->databaseName,
+        ]);
+    }
+
+    /**
+     * Make sure we have the mongo extension available and drop the test database just in case
+     */
+    public function setUp() {
+        if (!class_exists('MongoDB\Client')) {
+            $this->markTestSkipped('pecl/mongodb >= 1.1.3 is required to run this test');
+        }
+
+        $client = new MongoClient();
+        $client->dropDatabase($this->databaseName);
+
+        parent::setUp();
+    }
+
+    /**
+     * Drop the test database after each test
+     */
+    public function tearDown() {
+        if (class_exists('MongoDB\Client')) {
+            $client = new MongoClient();
+            $client->dropDatabase($this->databaseName);
+        }
+
+        parent::tearDown();
+    }
+}